fix: cast bigint correctly

+++ apps/ensapi/src/graphql-api/lib/find-domains/find-domains.ts
@@ -272,7 +272,8 @@ export function cursorFilter(
   // NOTE: Drizzle 0.41 doesn't support gt/lt with tuple arrays, so we use raw SQL
   // NOTE: explicit cast required — Postgres can't infer parameter types in tuple comparisons
   const op = useGreaterThan ? ">" : "<";
-  const value = cursor.by === "NAME" ? sql`${cursor.value}::text` : sql`${cursor.value}::bigint`;
+  const value =
+    cursor.by === "NAME" ? sql`${cursor.value}::text` : sql`${cursor.value}::numeric(78,0)`;
   return sql`(${orderColumn}, ${domains.id}) ${sql.raw(op)} (${value}, ${cursor.id})`;
 }
